# docs/openapi.yaml
openapi: 3.1.0
info:
  title: POPROG Knowledge Base API
  version: v1
  description: API для работы с публикациями, студенческими работами, загрузкой документов и полнотекстовым поиском.
...
paths:
  /api/assistant/chat:
    post:
      tags:
        - ИИ-ассистент
      summary: Отправить сообщение в чат с ИИ-ассистентом
      description: Принимает новые сообщения для диалога, при наличии chatId подмешивает сохраненную историю, отправляет запрос в GigaChat и возвращает ответ ассистента.
      operationId: assistantChat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiAssistantChatRequest"
      responses:
        "200":
          description: Ответ ИИ-ассистента успешно получен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiAssistantChatResponse"
        "400":
          description: Некорректные входные данные
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Диалог не найден
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "503":
          description: Интеграция с GigaChat временно недоступна
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /api/assistant/chats/{chatId}/messages:
    get:
      tags:
        - ИИ-ассистент
      summary: Получить историю диалога с ИИ-ассистентом
      description: Возвращает сохраненную историю сообщений указанного диалога в хронологическом порядке.
      operationId: getAssistantChatHistory
      parameters:
        - name: chatId
          in: path
          required: true
          description: Идентификатор диалога
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: История диалога успешно получена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatHistoryResponse"
        "404":
          description: Диалог не найден
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
...
components:
...
  schemas:
    AiAssistantChatRequest:
      type: object
      required:
        - messages
      properties:
        chatId:
          type:
            - string
            - "null"
          format: uuid
        messages:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/AiAssistantChatMessageRequest"
    AiAssistantChatMessageRequest:
      type: object
      required:
        - role
        - content
      properties:
        role:
          type: string
          enum:
            - system
            - user
            - assistant
        content:
          type: string
          minLength: 1
    AiAssistantChatResponse:
      type: object
      required:
        - chatId
        - content
        - model
      properties:
        chatId:
          type: string
          format: uuid
        content:
          type: string
        model:
          type: string
        finishReason:
          type:
            - string
            - "null"
        promptTokens:
          type:
            - integer
            - "null"
        completionTokens:
          type:
            - integer
            - "null"
        totalTokens:
          type:
            - integer
            - "null"
    ChatHistoryResponse:
      type: object
      required:
        - chatId
        - messages
      properties:
        chatId:
          type: string
          format: uuid
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatHistoryMessageResponse"
    ChatHistoryMessageResponse:
      type: object
      required:
        - id
        - role
        - content
        - createdAt
      properties:
        id:
          type: integer
          format: int64
        role:
          type: string
          enum:
            - system
            - user
            - assistant
        content:
          type: string
        createdAt:
          type: string
          format: date-time
...
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri